feat: Integrate with ratelimit module
This commit integrates the Defender module with the caddy-ratelimit
module to enable rate limiting based on IP ranges.  Defender now
marks requests from specified IP ranges with a header, and the
ratelimit module applies rate limiting only to requests with that
header.  This allows for flexible rate limiting based on IP ranges
or other criteria.

fixes #23